Restructured project
Assets were moved to a directory within the project root titled
`assets`.

Audio files are in a subdirectory titled `audio` which
contains two subdirectories itself, `final` and `split`. Final
is the final audio to be used within the game.  Split contains
individual tracks with lines from the voice actors.

Assets targeting print are within a subdirectory of `assets`
titled `print`.  These include the booklet and the board.